requirements.txt
Django>=3.2
djangorestframework>=3.12.2
psycopg2>=2.8.6
pygments>=2.9.0
djangorestframework-simplejwt>=4.7.2